@@ -4,12 +4,12 @@ function Footer() {
   return (
     <>
       <footer id="botFooter">
-        <a href="https://github.com/Ecole7853" id="github" target="_blank"></a>
         <a
           href="https://linkedin.com/in/ecole7853/"
           id="linkedin"
           target="_blank"
-        ></a>
+        ></a>{" "}
+        <a href="https://github.com/Ecole7853" id="github" target="_blank"></a>
         <a
           href="https://www.facebook.com/eric.cole.980"
           id="facebook"
